# controller/doc_response_models.go
package controller

import "threadnest/models"

// 专门用来放接口文档用到的model
// 因为我们的接口文档返回的数据格式是一致的，但是具体的data类型不一致

// _ResponsePostList 帖子列表接口响应数据
type _ResponsePostList struct {
	Code    ResCode        `json:"code"` // 业务响应状态码
	Message string         `json:"msg"`  // 提示信息
	Data    []*models.Post `json:"data"` // 数据
}

// _ResponsePostDetailList 增强帖子列表接口响应数据。
type _ResponsePostDetailList struct {
	Code    ResCode                 `json:"code"`
	Message string                  `json:"msg"`
	Data    []*models.ApiPostDetail `json:"data"`
}